rename portability-client-types to portability-types-client for consistency
