+++ portability-types-client/src/main/java/org/dataportabilityproject/types/client/package-info.java
@@ -16,4 +16,4 @@
 /**
  * Defines the data portability client model objects.  Model objects represent the system data resources a client has access to.
  */
-package org.dataportabilityproject.client.types;
+package org.dataportabilityproject.types.client;
