+++ portability-types-client/src/main/java/org/dataportabilityproject/types/client/providers/RegisteredServiceProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.client.types.providers;
+package org.dataportabilityproject.types.client.providers;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
